:sparkles: feat : [Cpp02] error hanling when floating point excpt

+++ cpp02/ex01/src/main.cpp
@@ -21,7 +21,7 @@ int main(void) {
 
   a = Fixed(1234.4321f);
 
-  std::cout << "a is " << a.toFloat() << std::endl;
+  std::cout << "a is " << a << std::endl;
   std::cout << "b is " << b << std::endl;
   std::cout << "c is " << c << std::endl;
   std::cout << "d is " << d << std::endl;
